src/He.PipelineAssessment.UI.Tests/Features/Ammendment/LoadAmmendmentCheckYourAnswers/LoadAmmendmentCheckYourAnswersRequestHandlerTests.cs
using AutoFixture.Xunit2;
using He.PipelineAssessment.Infrastructure.Repository;
using He.PipelineAssessment.Models;
using He.PipelineAssessment.Tests.Common;
using He.PipelineAssessment.UI.Features.Ammendment.LoadAmmendmentCheckYourAnswers;
using He.PipelineAssessment.UI.Features.Ammendment.SubmitAmmendment;
using He.PipelineAssessment.UI.Features.Intervention;
using Moq;
using Xunit;

namespace He.PipelineAssessment.UI.Tests.Features.Ammendment.LoadAmmendmentCheckYourAnswers
{
    public class LoadAmmendmentCheckYourAnswersRequestHandlerTests
    {
        [Theory]
        [AutoMoqData]
        public async Task Handle_ShouldError_GivenInterventionNotFound(
                   LoadAmmendmentCheckYourAnswersRequest request,
                   LoadAmmendmentCheckYourAnswersRequestHandler sut)
        {
            //Act
            var ex = await Assert.ThrowsAsync<ApplicationException>(() => sut.Handle(request, CancellationToken.None));

            //Assert
            Assert.Equal($"Unable to load ammendment check your answers. InterventionId: {request.InterventionId}", ex.Message);
        }

        [Theory]
        [AutoMoqData]
        public async Task Handle_ShouldReturn(
            [Frozen] Mock<IAssessmentRepository> repository,
            [Frozen] Mock<IAssessmentInterventionMapper> mapper,
            AssessmentInterventionCommand command,
            AssessmentIntervention intervention,
            LoadAmmendmentCheckYourAnswersRequest request,
            LoadAmmendmentCheckYourAnswersRequestHandler sut)
        {
            //Arrange
            repository.Setup(x => x.GetAssessmentIntervention(request.InterventionId))
                .ReturnsAsync(intervention);

            mapper.Setup(x => x.AssessmentInterventionCommandFromAssessmentIntervention(intervention))
                .Returns(command);

            //Act
            var result = await sut.Handle(request, CancellationToken.None);

            //Assert
            Assert.NotNull(result);
            Assert.IsType<SubmitAmmendmentCommand>(result);
        }
    }
}
